+++ src/components/EventChart/Cards/PowerSavingMode.tsx
@@ -7,10 +7,9 @@
 import React from 'react';
 import { useSelector } from 'react-redux';
 
-import { DocumentationKeys } from '../../../features/tracingEvents/at';
 import { getPowerSavingMode } from '../../../features/tracingEvents/dashboardSlice';
 import { PowerSavingModeValues } from '../../../features/tracingEvents/types';
-import DashboardCard from './DashboardCard';
+import DashboardCard, { DashboardCardFields } from './DashboardCard';
 
 const formatPSMValuesToString = (
     values: PowerSavingModeValues | undefined
@@ -19,6 +18,9 @@ const formatPSMValuesToString = (
         return 'Unknown';
     }
 
+    if (values.bitmask === '11100000') {
+        return `Deactivated = ${values.bitmask}`;
+    }
     if (values.unit && values.value) {
         return `${values.value} ${values.unit} = ${values.bitmask}`;
     }
@@ -31,14 +33,14 @@ export default () => {
         granted: undefined,
     };
 
-    const fields = {
+    const fields: DashboardCardFields = {
         'Requested Periodic TAU (T3412 extended)': {
             value: formatPSMValuesToString(requested?.T3412Extended),
-            commands: ['AT+CPSMS'] as DocumentationKeys[],
+            commands: ['AT+CPSMS'] as const,
         },
         'Requested Active Timer (T3324)': {
             value: formatPSMValuesToString(requested?.T3324),
-            commands: ['AT+CPSMS'] as DocumentationKeys[],
+            commands: ['AT+CPSMS'] as const,
         },
 
         // Could be displayed if user toggles an Advanced option?
@@ -47,15 +49,22 @@ export default () => {
         // 'T3402 Extended': requested.T3402_extended ?? 'Unknown',
 
         // GRANTED VALUES
+        'Power Saving Mode State': {
+            value: granted?.state?.toUpperCase() ?? 'OFF',
+            commands: ['AT+CEREG', 'AT%XMONITOR'],
+        },
         'Granted Periodic TAU (T3412 extended)': {
             value: formatPSMValuesToString(granted?.T3412Extended),
-            commands: [],
+            commands: ['AT+CEREG', 'AT%XMONITOR'],
         },
         'Granted Active Timer (T3324)': {
             value: formatPSMValuesToString(granted?.T3324),
-            commands: [],
+            commands: ['AT+CEREG', 'AT%XMONITOR'],
         },
-        // T3412: requested.T3412 ?? undefined,
+        'Granted Periodic TAU (T3412 / legacy)': {
+            value: formatPSMValuesToString(granted?.T3412),
+            commands: ['AT+CEREG', 'AT%XMONITOR'],
+        },
     };
 
     return (