+++ src/components/EventChart/Cards/Statistics.tsx
@@ -7,43 +7,43 @@
 import React from 'react';
 import { useSelector } from 'react-redux';
 
-import { DocumentationKeys } from '../../../features/tracingEvents/at';
+import { ATCommands } from '../../../features/tracingEvents/at';
 import { getDashboardState } from '../../../features/tracingEvents/dashboardSlice';
-import DashboardCard from './DashboardCard';
+import DashboardCard, { DashboardCardFields } from './DashboardCard';
 
 export default () => {
     const { connStat } = useSelector(getDashboardState);
 
-    const fields = {
+    const fields: DashboardCardFields = {
         'Collecting data': {
             value: parseCollecting(connStat?.collecting),
-            commands: ['AT%XCONNSTAT'] as DocumentationKeys[],
+            commands: ['AT%XCONNSTAT'] as ATCommands[],
         },
         'Successfull SMS Tx': {
             value: connStat?.smsTX ?? 'Unknown',
-            commands: ['AT%XCONNSTAT'] as DocumentationKeys[],
+            commands: ['AT%XCONNSTAT'] as ATCommands[],
         },
         'Successfull SMS Rx': {
             value: connStat?.smsRX ?? 'Unknown',
-            commands: ['AT%XCONNSTAT'] as DocumentationKeys[],
+            commands: ['AT%XCONNSTAT'] as ATCommands[],
         },
         'Data Transmitted': {
             value: connStat?.dataTX ? `${connStat?.dataTX} kB` : 'Unknown',
-            commands: ['AT%XCONNSTAT'] as DocumentationKeys[],
+            commands: ['AT%XCONNSTAT'] as ATCommands[],
         },
         'Data Recieved': {
             value: connStat?.dataRX ? `${connStat?.dataRX} kB` : 'Unknown',
-            commands: ['AT%XCONNSTAT'] as DocumentationKeys[],
+            commands: ['AT%XCONNSTAT'] as ATCommands[],
         },
         'Max Packet Size Tx or Rx': {
             value: connStat?.packetMax ? `${connStat.packetMax} kB` : 'Unknown',
-            commands: ['AT%XCONNSTAT'] as DocumentationKeys[],
+            commands: ['AT%XCONNSTAT'] as ATCommands[],
         },
         'Average Packet Size': {
             value: connStat?.packetAverage
                 ? `${connStat.packetAverage} kB`
                 : 'Unknown',
-            commands: ['AT%XCONNSTAT'] as DocumentationKeys[],
+            commands: ['AT%XCONNSTAT'] as ATCommands[],
         },
     };