+++ src/features/powerEstimation/powerEstimationSlice.ts
@@ -9,6 +9,7 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
 
 import type { RootState } from '../../appReducer';
+import { parseTAUByteToSeconds, TAU_TYPES } from '../../utils/powerSavingMode';
 import type { OnlinePowerEstimatorParams } from './onlinePowerEstimator';
 
 interface PowerEstimationState {
@@ -17,11 +18,6 @@ interface PowerEstimationState {
     renderedHtml: string | null;
     hasError: boolean;
     loading: boolean;
-}
-
-enum TAU_TYPES {
-    SLEEP_INTERVAL = 0,
-    ACTIVE_TIMER = 1,
 }
 
 const defaultPowerEstimationParameters: OnlinePowerEstimatorParams = {
@@ -165,42 +161,3 @@ export const {
 } = powerEstimationSlice.actions;
 
 export default powerEstimationSlice.reducer;
-
-const TAU_DEFAULT_VALUE = 0;
-const TAU_SLEEP_INTERVAL_BASE_VALUES: { [index: string]: number } = {
-    '000': 600,
-    '001': 3600,
-    '010': 36000,
-    '011': 2,
-    '100': 30,
-    '101': 60,
-};
-const TAU_ACTIVE_TIMER_BASE_VALUES: { [index: string]: number } = {
-    '000': 2,
-    '001': 60,
-};
-
-const parseTAUByteToSeconds = (byteString: string, type: TAU_TYPES) => {
-    const byteArray = [...byteString.trim()];
-
-    if (byteArray.length !== 8) return TAU_DEFAULT_VALUE;
-
-    const TAU_BASE_VALUE = byteArray.slice(0, 3).join('');
-    const TAU_BINARY_VALUE = Number.parseInt(byteArray.slice(3).join(''), 2);
-
-    if (
-        type === TAU_TYPES.SLEEP_INTERVAL &&
-        TAU_BASE_VALUE in TAU_SLEEP_INTERVAL_BASE_VALUES
-    )
-        return (
-            TAU_SLEEP_INTERVAL_BASE_VALUES[TAU_BASE_VALUE] * TAU_BINARY_VALUE
-        );
-
-    if (
-        type === TAU_TYPES.ACTIVE_TIMER &&
-        TAU_BASE_VALUE in TAU_ACTIVE_TIMER_BASE_VALUES
-    )
-        return TAU_ACTIVE_TIMER_BASE_VALUES[TAU_BASE_VALUE] * TAU_BINARY_VALUE;
-
-    return TAU_DEFAULT_VALUE;
-};