+++ src/features/tracingEvents/at/commandProcessors/dataProfile.ts
@@ -19,7 +19,7 @@ export type PowerLevel = keyof typeof PowerLevel;
 
 let requestedDataProfile: PowerLevel | undefined;
 
-export const processor: Processor = {
+export const processor: Processor<'%XDATAPRFL'> = {
     command: '%XDATAPRFL',
     documentation:
         'https://infocenter.nordicsemi.com/topic/ref_at_commands/REF/at_commands/mob_termination_ctrl_status/xdataprfl.html',