+++ src/features/tracingEvents/at/commandProcessors/evaluatingConnectionParameters.ts
@@ -14,7 +14,7 @@ import {
 import type { Processor } from '..';
 import { getParametersFromResponse } from '../utils';
 
-export const processor: Processor = {
+export const processor: Processor<'%CONEVAL'> = {
     command: '%CONEVAL',
     documentation:
         'https://infocenter.nordicsemi.com/topic/ref_at_commands/REF/at_commands/mob_termination_ctrl_status/coneval.html',